pyproject.toml
[project]
name = "gymme"
version = "0.1.0"
description = "百丽宫中关村羽毛球捡漏王"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hishel>=0.1.2",
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
    "rich>=14.0.0",
    "setuptools>=80.9.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["gymme", "gymme.*"]

[tool.ruff]
line-length = 120
